backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "MekaNavyasri",email = "navyasrimeka9@gmail.com"}
]

requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "qdrant-client (>=1.14.3,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "openai (>=1.93.0,<2.0.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
